// admin/src/types/RootTypes.tsx
export interface AdminTypes {
  createdAt: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  role: string;
  updatedAt: string;
  __v: number;
  _id: string;
}

export interface PartnerTypes {
  createdAt: string; 
  title: string;
  description: string;
  images: string[];
  link: string;
  updatedAt: string;
  __v: number;
  _id: string;
}

export interface ContactTypes {
  createdAt: string;
  name: string;
  email: string;
  message: string;
  updatedAt: string;
  __v: number;
  _id: string;
}


export interface CarouselTypes {
  createdAt: string;
  title: string;
  description: string;
  images: string[];
  link: string;
  updatedAt: string;
  __v: number;
  _id: string;
}